+++ src/com/adibrata/smartdealer/model/AssetServiceMaster.java
@@ -1,7 +1,8 @@
 
 package com.adibrata.smartdealer.model;
-// Generated Aug 20, 2015 1:54:39 PM by Hibernate Tools 4.3.1
+// Generated Aug 21, 2015 12:01:27 PM by Hibernate Tools 4.3.1
 
+import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.Entity; import org.hibernate.annotations.Cache; import org.hibernate.annotations.CacheConcurrencyStrategy; import javax.persistence.Cacheable;
@@ -25,7 +26,7 @@ public class AssetServiceMaster implements java.io.Serializable
 		private Partner partner;
 		private String serviceCode;
 		private String serviceName;
-		private Double priceMaximum;
+		private BigDecimal priceMaximum;
 		private String assetType;
 		private Date dtmUpd;
 		private String usrUpd;
@@ -41,7 +42,7 @@ public class AssetServiceMaster implements java.io.Serializable
 				this.id = id;
 			}
 			
-		public AssetServiceMaster(long id, Partner partner, String serviceCode, String serviceName, Double priceMaximum, String assetType, Date dtmUpd, String usrUpd, Date dtmCrt, String usrCrt)
+		public AssetServiceMaster(long id, Partner partner, String serviceCode, String serviceName, BigDecimal priceMaximum, String assetType, Date dtmUpd, String usrUpd, Date dtmCrt, String usrCrt)
 			{
 				this.id = id;
 				this.partner = partner;
@@ -103,12 +104,12 @@ public class AssetServiceMaster implements java.io.Serializable
 			}
 			
 		@Column(name = "PriceMaximum", precision = 17)
-		public Double getPriceMaximum()
+		public BigDecimal getPriceMaximum()
 			{
 				return this.priceMaximum;
 			}
 			
-		public void setPriceMaximum(Double priceMaximum)
+		public void setPriceMaximum(BigDecimal priceMaximum)
 			{
 				this.priceMaximum = priceMaximum;
 			}
