+++ src/com/adibrata/smartdealer/model/PersCustCcinfo.java
@@ -1,7 +1,8 @@
 
 package com.adibrata.smartdealer.model;
-// Generated Aug 20, 2015 1:54:39 PM by Hibernate Tools 4.3.1
+// Generated Aug 21, 2015 12:01:27 PM by Hibernate Tools 4.3.1
 
+import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.Column;
 import javax.persistence.Entity; import org.hibernate.annotations.Cache; import org.hibernate.annotations.CacheConcurrencyStrategy; import javax.persistence.Cacheable;
@@ -28,7 +29,7 @@ public class PersCustCcinfo implements java.io.Serializable
 		private String creditCardType;
 		private short creditCardPeriodYears;
 		private short creditCardPeriodMonths;
-		private Double creditCardCreditLimit;
+		private BigDecimal creditCardCreditLimit;
 		private Date dtmUpd;
 		private String usrUpd;
 		private Date dtmCrt;
@@ -38,7 +39,7 @@ public class PersCustCcinfo implements java.io.Serializable
 			{
 			}
 			
-		public PersCustCcinfo(long id, String creditCardNo, String creditCardName, String creditCardType, short creditCardPeriodYears, short creditCardPeriodMonths, Double creditCardCreditLimit)
+		public PersCustCcinfo(long id, String creditCardNo, String creditCardName, String creditCardType, short creditCardPeriodYears, short creditCardPeriodMonths, BigDecimal creditCardCreditLimit)
 			{
 				this.id = id;
 				this.creditCardNo = creditCardNo;
@@ -49,7 +50,7 @@ public class PersCustCcinfo implements java.io.Serializable
 				this.creditCardCreditLimit = creditCardCreditLimit;
 			}
 			
-		public PersCustCcinfo(long id, Customer customer, String creditCardNo, String creditCardName, String creditCardType, short creditCardPeriodYears, short creditCardPeriodMonths, Double creditCardCreditLimit, Date dtmUpd, String usrUpd,
+		public PersCustCcinfo(long id, Customer customer, String creditCardNo, String creditCardName, String creditCardType, short creditCardPeriodYears, short creditCardPeriodMonths, BigDecimal creditCardCreditLimit, Date dtmUpd, String usrUpd,
 		        Date dtmCrt, String usrCrt)
 			{
 				this.id = id;
@@ -147,12 +148,12 @@ public class PersCustCcinfo implements java.io.Serializable
 			}
 			
 		@Column(name = "CreditCardCreditLimit", nullable = false, precision = 17)
-		public Double getCreditCardCreditLimit()
+		public BigDecimal getCreditCardCreditLimit()
 			{
 				return this.creditCardCreditLimit;
 			}
 			
-		public void setCreditCardCreditLimit(Double creditCardCreditLimit)
+		public void setCreditCardCreditLimit(BigDecimal creditCardCreditLimit)
 			{
 				this.creditCardCreditLimit = creditCardCreditLimit;
 			}
